max-value: Create a solution
max-value.js exports function which accepts an array of types and a
capacity and returns the maximum possibly value. Infinity is returned if
a type has a 0 weight and positive value. Uses unbounded knapsack
algorithm with a time complexity of O(n * k) where n is the length of
the types array and k is the capacity.

about.md may require modification to better represent the solution.